Part1 npu 1.2

+++ npu/conv_npu.py
@@ -200,7 +200,7 @@ def conv2d(X, W, bias):
                     c_out_tile_start = c_out_tile_i*c_out_pmax
                     c_out_tile_end = c_out_tile_start+c_out_pmax
                     nl.store(X_out[b_i,c_out_tile_start:c_out_tile_end,output_h_start+out_h_i ],
-                        Output_row, dtype=X_out.dtype)
+                        Output_row)
                       
     return X_out
